+++ HISTORY.md
@@ -1,3 +1,22 @@
+# 0.3.0 / 2015.03.18
+
+  * Add parsing from object
+  * Add alpha to Color
+  * Add output to object and css string
+  * Add rgba, hsl(a), hsv(a) models support, fix #2
+  * Add parsing from array
+  * Add parsing color name, fix #4
+  * Add output to array
+  * Fix #3 parsing from css color string
+  * Fix parsing hex
+  * Fix convert to color name
+  * Fix hsl and hsv conversions
+  * Fix negative deg values
+  * Fix fractional values
+  * Extract fmt
+  * Remove string parsing via ccc methods
+  * Update test
+
 # 0.2.0 / 2015.02.19
 
   * Add average calculation #1